ADDED LOGIN LOGIC BUT INCOMPLETE(WITHOUT DATABASE)
