+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.MicroBloggingApp"
         tools:targetApi="31">
         <activity
-            android:name=".LOGINACTIVITY"
+            android:name=".LoginActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
